pip-requirements.txt
apt-xapian-index==0.45
behave==1.2.5
chardet==2.0.1
Cheetah==2.4.4
cloud-init==0.7.5
colorama==0.2.5
configobj==4.7.2
coverage==3.7.1
Django==1.8.2
django-behave==0.1.1
django-bower==5.0.1
django-cors-headers==1.1.0
django-filter==0.10.0
django-multiselectfield==0.1.3
django-redis-sessions==0.4.0
django-registration==1.0
django-taggit==0.12
-e git+https://github.com/jrief/django-websocket-redis@5ff598dd5f3f37067011e7b6d9cb7c7fbb019aba#egg=django_websocket_redis-dev
djangorestframework==3.1.3
djangorestframework-jwt==1.6.0
enum34==1.0.4
gevent==0.13.6
gevent-socketio==0.2.1
gevent-websocket==0.2.3
greenlet==0.4.2
html5lib==0.999
httplib2==0.8
jsonpatch==1.3
jsonpointer==1.0
keyring==3.5
Landscape-Client==14.1
launchpadlib==1.10.2
lazr.restfulclient==0.13.3
lazr.uri==1.0.3
libsaas==0.4
Markdown==2.6.2
mock==1.0.1
MockMockMock==0.5.1
nose==1.2.1
numpy==1.8.2
oauth==1.0.1
PAM==0.4.2
parse==1.6.6
parse-type==0.3.4
phonenumbers==6.1.0
Pillow==2.4.0
poster==0.8.1
prettytable==0.7.2
pycrypto==2.6.1
pycurl==7.19.3
pygobject==3.12.0
PyJWT==1.3.0
pyOpenSSL==0.13
pyserial==2.6
python-apt==0.9.3.5
python-dateutil==2.2
PyYAML==3.10
redis==2.10.1
requests==1.1.0
SecretStorage==2.0.0
simplejson==3.3.1
six==1.7.2
sphinx-me==0.2.1
testresources==0.2.7
Twisted-Core==13.2.0
Twisted-Names==13.2.0
Twisted-Web==13.2.0
Unirest==1.1.7
uritemplate.py==0.3.0
urllib3==1.7.1
virtualenv==1.11.6
wadllib==1.3.2
Wand==0.3.7
wheel==0.24.0
zope.interface==4.0.5
psycopg2
unipath
django-polymorphic